Rename view -> component
