+++ app/src/main/java/com/litechmeg/sabocale/component/adapter/AttendanceListArrayAdapter.java
@@ -1,4 +1,4 @@
-package com.litechmeg.sabocale.view.adapter;
+package com.litechmeg.sabocale.component.adapter;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
